Fix: mark done if/only if we have an error
If a tx is_done, we should not return done. We need this tx to go around
the spin() gauntlet one more time so that _prepare_tansactions()'s
gets the chance to call disconnect() and make the whole connection
cache thing work properly.

Without this, keep-alive connections would not work.

Signed-off-by: Pedro Melo <melo@simplicidade.org>